+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changes
 * [UI]: Fixed bug on users listing page where if a user has never logged into the system there last login date showed `Dec 31 1969`,
 * [Developer]: Removed UI dependency for `bootstrap-sass`.
 * [Developer]: Updated to React v17.0.1, Babel v7.12.1, and Ant Design v4.7.3.
+* [UI]: Updated FastQC page to use Ant Design.
 
 20.05 to 20.09
 --------------